requirements.txt
langchain
wikipedia
langchain-community
groq
python-dotenv
pydantic
langchain-groq
duckduckgo-search
streamlit